refactor: 약관 Drawer 분리 관리
